replacing react render with redux example
